requirements.txt
SQLAlchemy~=2.0.34
loguru~=0.7.2
fastapi~=0.114.0
starlette~=0.38.5
pydantic~=2.9.1
uvicorn~=0.30.6
python-dotenv~=1.0.1
redis~=5.0.8
celery~=5.4.0
aiosmtplib~=3.0.2
asyncpg
pydantic_core
jinja2
python-multipart
pydantic[email]
PyJWT~=2.9.0
alembic==1.13.3
gunicorn